Convert to dotnet6.0
