--- Multicast/LocalTestRun.testrunconfig
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<TestRunConfiguration name="Local Test Run" id="e3ece893-a0aa-49ab-99fa-48b6ecfa8aff" xmlns="http://microsoft.com/schemas/VisualStudio/TeamTest/2010">
-  <Description>This is a default test run configuration for a local test run.</Description>
-  <TestTypeSpecific />
-</TestRunConfiguration>
