--- Multicast/Multicast-vs2010.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 10
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{7C1F24F3-253B-45E6-981E-791778F8AD8F}"
-	ProjectSection(SolutionItems) = preProject
-		localtestrun.testrunconfig = localtestrun.testrunconfig
-		multicast.vsmdi = multicast.vsmdi
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multicast", "Multicast\Multicast.csproj", "{CA88F6EA-3279-4FB1-8850-856EB511F300}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multicast.Tests", "Multicast.Tests\Multicast.Tests.csproj", "{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}"
-EndProject
-Global
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = Multicast.vsmdi
-	EndGlobalSection
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|x86.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|x86.Build.0 = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Any CPU.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Mixed Platforms.Build.0 = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|x86.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|x86.Build.0 = Release|x86
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{BDEEC51D-9DC4-4CFD-A265-6C5995BB1C46}.Release|x86.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
