--- Multicast/Multicast.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multicast", "Multicast\Multicast.csproj", "{CA88F6EA-3279-4FB1-8850-856EB511F300}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multicast.Tests", "Multicast.Tests\Multicast.Tests.csproj", "{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{148963AD-96FE-466C-B949-561AF6547E9A}"
-	ProjectSection(SolutionItems) = preProject
-		LocalTestRun.testrunconfig = LocalTestRun.testrunconfig
-		Multicast.vsmdi = Multicast.vsmdi
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = Multicast.vsmdi
-	EndGlobalSection
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|x86.ActiveCfg = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Debug|x86.Build.0 = Debug|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Any CPU.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|Mixed Platforms.Build.0 = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|x86.ActiveCfg = Release|x86
-		{CA88F6EA-3279-4FB1-8850-856EB511F300}.Release|x86.Build.0 = Release|x86
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{8CD6AE80-E6F0-470C-B9BB-EF75809B0030}.Release|x86.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
