--- Multicast/Multicast/IMulticastBroadcaster.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Multicast
-{
-    public interface IMulticastBroadcaster : IDisposable
-    {
-        MulticastSettings Settings { get; }
-
-        bool IsBound { get; }
-
-        void Broadcast(byte[] data);
-    }
-}
