+++ src/Multicast/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace Multicast;
+
+public class Class1
+{
+}
