+++ src/Multicast/IMulticastBroadcaster.cs
@@ -1,0 +1,11 @@
+﻿namespace Multicast
+{
+    public interface IMulticastBroadcaster : IDisposable
+    {
+        MulticastSettings Settings { get; }
+
+        bool IsBound { get; }
+
+        void Broadcast(byte[] data);
+    }
+}
